requirements.txt
astroid==2.15.8
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
dill==0.3.7
dnspython==2.4.2
exceptiongroup==1.1.3
idna==3.4
iniconfig==2.0.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==1.26.0
packaging==23.2
pandas==2.1.1
platformdirs==3.10.0
pluggy==1.3.0
pyaes==1.6.1
pycodestyle==2.11.0
pyflakes==3.1.0
pylint==2.17.6
pymongo==4.5.0
Pyrogram==2.0.106
PySocks==1.7.1
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
TgCrypto==1.2.5
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
websocket-client==0.56.0
wrapt==1.15.0
